source/requirements.txt
# Core libraries
numpy
pandas
scikit_learn
tqdm
sacrebleu

# Machine learning frameworks
torch
sentence_transformers==3.3.1
transformers==4.46.3

# Retrieval and indexing
faiss-cpu
rank_bm25==0.2.2

# Langchain and OpenAI integration
langchain
huggingface-hub
bitsandbytes
nltk
accelerate==1.1.1
langchain_community
chromadb>=0.5.21
tiktoken
openai==1.57.0

# Jupyter notebook tools
notebook
jupyter_contrib_nbextensions
jupyter_nbextensions_configurator

# Python package tools
pip>=22.0
wheel>=0.38.4
setuptools>=65.5.1